KYLIN-1924 Region server metrics: replace int type for long type for scanned row count

+++ core-cube/src/main/java/org/apache/kylin/gridtable/GTAggregateScanner.java
@@ -119,7 +119,7 @@ public class GTAggregateScanner implements IGTScanner {
     }
 
     @Override
-    public int getScannedRowCount() {
+    public long getScannedRowCount() {
         return inputScanner.getScannedRowCount();
     }
